Split InputManagerPlugin into its own file
